[WebInterface] init trading_type_selector page
