+++ Services/Interfaces/web_interface/controllers/profiles.py
@@ -29,8 +29,7 @@ import octobot_services.interfaces.util as interfaces_util
 @login.login_required_when_activated
 def profiles_selector():
     reboot = flask.request.args.get("reboot", "false").lower() == "true"
-    show_nav_bar = flask.request.args.get("show_nav_bar", 'true').lower() != "false"
-    reboot_delay = 2
+    onboarding = flask.request.args.get("onboarding", 'false').lower() == "true"
     profiles = models.get_profiles()
     current_profile = models.get_current_profile()
     display_config = interfaces_util.get_edited_config()
@@ -58,11 +57,12 @@ def profiles_selector():
     display_intro = flask_util.BrowsingDataProvider.instance().get_and_unset_is_first_display(
         flask_util.BrowsingDataProvider.PROFILE_SELECTOR
     )
-    return_val = flask.render_template(
+    return flask.render_template(
         'profiles_selector.html',
-        show_nab_bar=show_nav_bar,
+        show_nab_bar=not onboarding,
+        onboarding=onboarding,
         read_only=True,
-        waiting_reboot=reboot,
+        reboot=reboot,
         display_intro=display_intro,
 
         current_logged_in_email=logged_in_email,
@@ -81,7 +81,3 @@ def profiles_selector():
 
         symbol_list=sorted(models.get_symbol_list(enabled_exchanges or config_exchanges)),
     )
-    if reboot and not models.is_rebooting():
-        # schedule reboot now that the page render has been computed
-        models.restart_bot(delay=reboot_delay)
-    return return_val
