+++ Services/Interfaces/web_interface/controllers/trading.py
@@ -18,6 +18,7 @@ import datetime
 import flask
 
 import octobot_services.interfaces.util as interfaces_util
+import octobot_commons.constants as commons_constants
 import tentacles.Services.Interfaces.web_interface as web_interface
 import tentacles.Services.Interfaces.web_interface.login as login
 import tentacles.Services.Interfaces.web_interface.models as models
@@ -98,6 +99,34 @@ def trades():
     return flask.render_template('trades.html',
                                  real_trades_history=real_trades_history,
                                  simulated_trades_history=simulated_trades_history)
+
+
+@web_interface.server_instance.route("/trading_type_selector")
+@login.login_required_when_activated
+def trading_type_selector():
+    reboot = flask.request.args.get("reboot", "false").lower() == "true"
+    onboarding = flask.request.args.get("onboarding", 'false').lower() == "true"
+    display_config = interfaces_util.get_edited_config()
+
+    return_val = flask.render_template(
+        'trading_type_selector.html',
+        show_nab_bar=not onboarding,
+        onboarding=onboarding,
+        waiting_reboot=reboot,
+
+        config_exchanges=display_config[commons_constants.CONFIG_EXCHANGES],
+        ccxt_tested_exchanges=models.get_tested_exchange_list(),
+        ccxt_simulated_tested_exchanges=models.get_simulated_exchange_list(),
+        ccxt_other_exchanges=sorted(models.get_other_exchange_list()),
+
+        config_trader=display_config[commons_constants.CONFIG_TRADER],
+        config_trader_simulator=display_config[commons_constants.CONFIG_SIMULATOR],
+    )
+    if reboot and not models.is_rebooting():
+        reboot_delay = 2
+        # schedule reboot now that the page render has been computed
+        models.restart_bot(delay=reboot_delay)
+    return return_val
 
 
 @web_interface.server_instance.context_processor
