+++ Services/Interfaces/web_interface/static/js/components/trading_type_selector.js
@@ -1,0 +1,28 @@
+/*
+ * Drakkar-Software OctoBot
+ * Copyright (c) Drakkar-Software, All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3.0 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library.
+ */
+
+$(document).ready(function() {
+    const onReconnected = () => {
+        const loader = $("#restart-loader");
+        if(loader.length){
+            // change current page when reconnected
+            window.location.href = loader.data("redirect-url");
+        }
+    }
+    registerReconnectedCallback(onReconnected);
+});
